Inline immutable strings and audit decomp quality + recover class fields across switch control flow

+++ decomp/ST.exe/functions/0041D350/decomp.c
@@ -16,7 +16,7 @@ int __thiscall FUN_0041d350(void *this,uint param_1)
   iVar1 = (int)*(short *)((int)this + 0x5b);
   iVar2 = (int)*(short *)((int)this + 0x5d);
   if ((((-1 < iVar1) && (iVar1 < (int)g_visibleClass_00802A88->field_0020)) && (-1 < iVar2)) &&
-     (((iVar2 < (int)g_visibleClass_00802A88->field_0024 && (param_1 < 8)) &&
+     (((iVar2 < g_visibleClass_00802A88->field_0024 && (param_1 < 8)) &&
       (*(int *)(&g_visibleClass_00802A88->field_0x74 + param_1 * 4) != 0)))) {
     return (uint)(*(char *)(g_visibleClass_00802A88->field_0020 * iVar2 +
                             *(int *)(&g_visibleClass_00802A88->field_0x74 + param_1 * 4) + iVar1) !=
