+++ decomp/ST.exe/functions/005F3940/decomp.c
@@ -63,14 +63,14 @@ int FUN_005f3940(int param_1,int param_2,int param_3)
     if (param_3 < 0) {
       return iVar3;
     }
-    if ((int)pVVar2->field_0030 <= param_3) {
+    if (pVVar2->field_0030 <= param_3) {
       return iVar3;
     }
     local_8 = (&DAT_0079aed0)[param_2] + local_8;
     if (local_8 < 0) {
       return iVar3;
     }
-    if ((int)pVVar2->field_0034 <= local_8) {
+    if (pVVar2->field_0034 <= local_8) {
       return iVar3;
     }
     if (pVVar2->field_004C == 0) {
