+++ decomp/ST.exe/functions/0060A4B0/decomp.c
@@ -67,9 +67,9 @@ undefined4 __fastcall FUN_0060a4b0(AnonShape_0060A4B0_7BF9A20C *param_1)
                 ((thunk_FUN_00558c00(g_visibleClass_00802A88,g_visibleClass_00802A88->field_010C,
                                      iVar5,iVar6,&local_1c,&local_20), local_8 < 0 ||
                  ((4 < local_8 || (local_1c < 0)))))) ||
-               (((int)pVVar4->field_0030 <= local_1c ||
+               ((pVVar4->field_0030 <= local_1c ||
                 ((((iVar6 = (&DAT_0079aed0)[local_8] + local_20, iVar6 < 0 ||
-                   ((int)pVVar4->field_0034 <= iVar6)) || (pVVar4->field_004C == 0)) ||
+                   (pVVar4->field_0034 <= iVar6)) || (pVVar4->field_004C == 0)) ||
                  (*(char *)(iVar6 * pVVar4->field_0030 + pVVar4->field_004C + local_1c) != '\0')))))
                ) {
               if (*(int *)(iVar7 + 4) == 0) {
