+++ decomp/ST.exe/functions/006226C0/decomp.c
@@ -47,9 +47,9 @@ uint __thiscall FUN_006226c0(void *this,uint param_1,int param_2)
             (uVar2 = thunk_FUN_00558c00(g_visibleClass_00802A88,g_visibleClass_00802A88->field_010C,
                                         (int)*(short *)((int)this + 0x47),
                                         (int)*(short *)((int)this + 0x49),&param_2,(int *)&local_8),
-            iVar3 < 0)) || (4 < iVar3)) || ((param_2 < 0 || ((int)pVVar1->field_0030 <= param_2))))
-         || ((uVar2 = (&DAT_0079aed0)[iVar3] + (int)local_8, (int)uVar2 < 0 ||
-             (((int)pVVar1->field_0034 <= (int)uVar2 || (pVVar1->field_004C == 0)))))) ||
+            iVar3 < 0)) || (4 < iVar3)) || ((param_2 < 0 || (pVVar1->field_0030 <= param_2)))) ||
+         ((uVar2 = (&DAT_0079aed0)[iVar3] + (int)local_8, (int)uVar2 < 0 ||
+          ((pVVar1->field_0034 <= (int)uVar2 || (pVVar1->field_004C == 0)))))) ||
         (uVar2 = (uint)*(byte *)(uVar2 * pVVar1->field_0030 + pVVar1->field_004C + param_2),
         uVar2 != 0)))) goto cf_common_exit_006227F9;
   }
