+++ decomp/ST.exe/functions/0062FEA0/decomp.c
@@ -37,14 +37,14 @@ int __thiscall FUN_0062fea0(void *this,int param_1,int param_2)
     if (param_2 < 0) {
       return iVar2;
     }
-    if ((int)pVVar1->field_0030 <= param_2) {
+    if (pVVar1->field_0030 <= param_2) {
       return iVar2;
     }
     iVar3 = (&DAT_0079aed0)[iVar3] + (int)local_8;
     if (iVar3 < 0) {
       return iVar2;
     }
-    if ((int)pVVar1->field_0034 <= iVar3) {
+    if (pVVar1->field_0034 <= iVar3) {
       return iVar2;
     }
     if (pVVar1->field_004C == 0) {
